requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
google-genai>=0.2.0
python-multipart>=0.0.6